requirements_current.txt
absl-py==2.3.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
anyio==4.11.0
archspec @ file:///croot/archspec_1709217642129/work
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-lru==2.0.5
attrs @ file:///croot/attrs_1695717823297/work
babel==2.17.0
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1718029820055/work
bleach==6.2.0
boltons @ file:///work/ci_py311/boltons_1677685195580/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1720453481653/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
chardet @ file:///work/ci_py311/chardet_1676830276092/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///croot/click_1698129812380/work
comm==0.2.3
conda @ file:///croot/conda_1715635703388/work
conda-build @ file:///croot/conda-build_1716991285767/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_index @ file:///croot/conda-index_1719338209492/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
contourpy==1.3.3
cramjam==2.11.0
cryptography @ file:///croot/cryptography_1714660666131/work
cycler==0.12.1
debugpy==1.8.17
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
distro @ file:///croot/distro_1714488253808/work
dnspython==2.6.1
docopt==0.6.2
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
expecttest==0.2.1
fastjsonschema==2.21.2
fastparquet==2024.11.0
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.60.1
fqdn==1.5.1
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2024.6.1
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hypothesis==6.108.4
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
ipykernel==7.0.1
ipython @ file:///croot/ipython_1718287989724/work
ipywidgets==8.1.7
isoduration==20.11.0
jedi @ file:///croot/jedi_1721058342488/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.5.2
json5==0.12.1
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema @ file:///croot/jsonschema_1699041609003/work
jsonschema-specifications @ file:///croot/jsonschema-specifications_1699032386549/work
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
kornia==0.7.4
kornia_rs==0.1.9
lazy_loader==0.4
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
lintrunner==0.12.5
llvmlite==0.45.1
Markdown==3.9
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.10.7
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
menuinst @ file:///croot/menuinst_1718132535249/work
mistune==3.1.4
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
more-itertools @ file:///croot/more-itertools_1700662129964/work
mpmath @ file:///croot/mpmath_1690848262763/work
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1720002482208/work
ninja==1.11.1.1
notebook_shim==0.2.4
numba==0.62.1
numpy==1.23.5
nvidia-ml-py3==7.352.0
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
optree==0.12.1
overrides==7.7.0
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.3.3
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
patsy==1.0.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
Pillow==9.4.0
pkginfo @ file:///croot/pkginfo_1715695984887/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
plotly==6.3.1
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
prometheus_client==0.23.1
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==6.33.0
psutil @ file:///work/ci_py311_2/psutil_1679337388738/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///croot/pygments_1684279966437/work
pykwalify==1.8.0
pynndescent==0.5.13
pyparsing==3.2.5
pyradiomics==3.0.1
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-json-logger==4.0.0
pytz @ file:///croot/pytz_1713974312559/work
PyWavelets==1.8.0
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==27.1.0
referencing @ file:///croot/referencing_1699012038513/work
requests @ file:///croot/requests_1721410876868/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py @ file:///croot/rpds-py_1698945930462/work
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
scikit-image==0.19.3
scikit-learn==1.6.1
scikit-posthocs==0.11.4
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
simpleitk==2.5.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve @ file:///croot/soupsieve_1696347547217/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
statsmodels==0.14.5
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tinycss2==1.4.0
torch==2.4.0
torch-staintools==1.0.4
torchaudio==2.4.0
torchelastic==0.2.2
torchvision==0.19.0
tornado==6.5.2
tqdm==4.65.2
traitlets @ file:///croot/traitlets_1718227057033/work
triton==3.0.0
truststore @ file:///croot/truststore_1695244293384/work
types-dataclasses==0.6.6
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1718912636303/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
zstandard @ file:///croot/zstandard_1714677652653/work
